mpesa-dashboard-backend/pyproject.toml
[tool.poetry]
name = "mpesa-csv-processor"
version = "0.1.0"
description = "A FastAPI backend for processing M-Pesa CSV uploads"
authors = ["Morgan <ngetichmorgan6@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.6.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"  # Async PostgreSQL
python-docx = "^0.8.11"  # DOCX processing
pymupdf = "^1.23.7"
python-multipart = "^0.0.20"




[tool.poetry.dev-dependencies]
black = "^24.2.0"
isort = "^5.12.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
